Moves the store implementation to a single file and shortens the delays for faster testing
